requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Django==3.0.3
django-filter==2.2.0
django-lockdown==3.0.0
djangorestframework==3.11.0
idna==2.8
oauthlib==3.1.0
pathspec==0.7.0
Pillow==7.0.0
psycopg2-binary==2.8.4
pytz==2018.5
regex==2020.2.20
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
toml==0.10.0
twython==3.7.0
typed-ast==1.4.1
urllib3==1.25.7
